api/2 project call now includes tags
